requirements.txt
gdown>=4.6.6
PyYaml
bardapi
unidecode
tqdm
openai
aiohttp
wandb
transformers
numpy
matplotlib
seaborn
